feat: Add credhub-setup container
This is a helper binary that will set up CF-level application security
groups to let the apps talk to CredHub (and the UAA authentication
required to use CredHub), so that the CredHub-related CATS tests will
pass.

For now, this hard-codes the host and port information for UAA; this
will need to be fixed eventually.

This only includes the actual source code; the build machinery will be
added in a subsequent commit.